Fix error 404 when clicking Home button
Modified the href link location for the Home button on carbonara.html,
french-toast.html, and veggie-tacos.html

Trying to go back to index.html by clicking the Home button when being
in one of the recipe pages exhibited the 404 error due to the href link
location being wrong. This behavior is now fixed.